function/requirements.txt
psycopg2-binary==2.8.6
scipy==1.5.3
Pillow==8.0.0
numpy==1.19.2
azure-storage-blob==12.5.0
azure-functions==1.4.0
